Organize models as feature sliced into ClinicManagement.Domain #3
