+++ ClinicManagement/src/Core/Fnunez.VeterinaryClinic.ClinicManagement.Domain/ClientAggregate/Enums/AnimalSex.cs
@@ -1,0 +1,7 @@
+namespace Fnunez.VeterinaryClinic.ClinicManagement.Domain.ClientAggregate.Enums;
+
+public enum AnimalSex
+{
+    Female,
+    Male
+}
