+++ ClinicManagement/src/Core/Fnunez.VeterinaryClinic.ClinicManagement.Domain/ClientAggregate/ValueObjects/AnimalType.cs
@@ -1,6 +1,6 @@
 using Fnunez.VeterinaryClinic.SharedKernel.Domain.Common;
 
-namespace Fnunez.VeterinaryClinic.ClinicManagement.Domain.ValueObjects;
+namespace Fnunez.VeterinaryClinic.ClinicManagement.Domain.ClientAggregate.ValueObjects;
 
 public class AnimalType : ValueObject
 {
