--- ClinicManagement/src/Core/Fnunez.VeterinaryClinic.ClinicManagement.Domain/Enums/AnimalSex.cs
@@ -1,7 +1,0 @@
-namespace Fnunez.VeterinaryClinic.ClinicManagement.Domain.Enums;
-
-public enum AnimalSex
-{
-    Female,
-    Male
-}
